Moves fizzbuzz  to service and consumes it in pipe
Testing a fizzbuzz service

Using the pipe was nice, but still isn't very
angular and will only work in templates. So we
put the logic in a service and test that service.
This gives us the chance to examine the angular
testbed in the service test but also the pipe
test since we test it in module.