Adds fizzbuzz pipe isolate test
Testing the fizzbuzz pipe in isolation

In step 3 we tested the service directly and
through the pipe, but an arguement could be made
the only thing we need to test is that the pipe
called the service. In step four we do this,
testing the pipe in isolation and exploring the
basics of spies.