Uses pokestuff in pokedex app
Putting it all together

This is the last step in the series. We are now
ready to put it all together. We add the PokeAPI
service to the app and the component to the app.

We make out inital request in the app.component
and push the output to our cards.